Add files via upload
This is my first ever project in C. I did it at the end of my first semester as a part of my course. It is a basketball database system. It has its limits with the number of teams that can enter, the number of players, etc. I would say things could have been done better in terms of the code written but as I was a newbie I did things with my understanding of the knowledge. If you have any comments regarding improving the lines of code in the project do leave your thoughts!